frontend-plugin-api: refactor to remove root modules
