+++ packages/frontend-plugin-api/src/schema/types.ts
@@ -14,20 +14,10 @@
  * limitations under the License.
  */
 
-export { coreExtensionData } from './coreExtensionData';
-export {
-  createExtension,
-  type CreateExtensionOptions,
-  type ExtensionDataBind,
-  type ExtensionDataValue,
-} from './createExtension';
-export {
-  createExtensionDataRef,
-  type ExtensionDataRef,
-} from './createExtensionDataRef';
-export {
-  createPlugin,
-  type BackstagePlugin,
-  type PluginOptions,
-} from './createPlugin';
-export type { AnyExtensionDataMap, Extension } from './types';
+import { JsonObject } from '@backstage/types';
+
+/** @public */
+export type PortableSchema<TOutput> = {
+  parse: (input: unknown) => TOutput;
+  schema: JsonObject;
+};
